+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * refactor: refactor `ObjectToObjectTransformer` for future optimization
 * fix: remove unalterable check in `ObjectProcessor` to allow transformers to
   handle the current target value in all cases
+* test: test lazy constructor
 
 ## 1.11.0